sct_deepseg_sc: implement lazy loading
Functions were moved to a scripting entry point in
 spinalcordtoolbox/deepseg_sc/


Former-commit-id: c3737e3140f4c69a070846bfb0a46de72e0e53bf [formerly 99f124891252972592dc94006e16bb636979f531 [formerly 62f7be496d5a629c504f84076266a64a10b8ba31]]
Former-commit-id: 49cf599cfc91e93b310b7c631c5812987fad3868
Former-commit-id: 81b6552b53f138b3cb73423b070f9e95a3474238
